+++ R/Misc_fun.R
@@ -35,13 +35,42 @@ matplot <- function(..., zero_line = FALSE) {
 
 #' @importFrom graphics hist legend
 hist.numeric <- function(x, ...) {
+  n <- length(x)
+  
+  # Remove NA's
   if (any(is.na(x))) {
     na_rate <- mean(is.na(x))
     x <- x[!is.na(x)]
+    legend_na <- paste0(round(100 * na_rate), "% NA's")
   } else {
-    na_rate <- 0
+    legend_na <- NULL
   }
-  if (all(!diff(signif(x, 3)))) {
+  
+  # Calculate skewness NA's already removed, n is the original length of x
+  skewness <- (sum((x - mean(x))^3)/n)/(sum((x - mean(x))^2)/n)^1.5
+  
+  # Remove outliers
+  if (!is.na(skewness)) {
+    if (skewness > 3) {
+      max_x_plot <- quantile(x, 0.95)
+      x <- x[x <= max_x_plot]
+      legend_skew1 <- paste0(round(100 * mean(max_x_plot/n)), "% > ", round(max_x_plot, 2))
+    } else {
+      legend_skew1 <- NULL
+    }
+    if (skewness < 3) {
+      min_x_plot <- quantile(x, 0.05)
+      x <- x[x >= min_x_plot]
+      legend_skew2 <- paste0(round(100 * mean(min_x_plot/n)), "% < ", round(min_x_plot, 2))
+    } else {
+      legend_skew2 <- NULL
+    }
+  } else {
+    legend_skew1 <- legend_skew2 <- NULL
+  }
+  
+  # Plot histogram
+  if (all(!diff(signif(x, 3)))) { # If all identical values
     x <- signif(x, 3)
     if (all(!x)) { # if x is all zeros
       breaks <- c(-0.1, 0.1)
@@ -54,11 +83,11 @@ hist.numeric <- function(x, ...) {
   } else {
     r <- graphics::hist.default(x, ...)
   }
-  if (na_rate > 0) {
-    legend("topright", 
-           paste0(round(100 * na_rate), "% NA's"), 
-           bty = "n",
-           text.col = "red")
+  
+  # Make legend
+  legend_text <- c(legend_na, legend_skew1, legend_skew2)
+  if (!is.null(legend_text)) {
+    legend("topright", legend = legend_text, bty = "n", text.col = "red")
   }
   invisible(r)
 }